pyproject.toml
[project]
name = "gesture-control-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "pyobjc-framework-cocoa>=12.0",
    "pyobjc-framework-quartz>=12.0",
    "pyyaml>=6.0.3",
]
